Drawing Board
Created a web page that includes an SVG area acting as a drawing canvas. When the user clicks and drags inside this area, a shape (for example, a line or rectangle) should be dynamically drawn following the mouse movement. The tool should respond to mouse events such as mousedown, mousemove, and mouseup to create and update the shapes in real time. Users should be able to draw multiple shapes by repeating the click-and-drag action. All drawing interactions is implemented using JavaScript without reloading the page.